Normalize mount targets (#4031)
Fixes #3733.

Co-authored by @ntorresalberto (thanks!)

This takes #3908 and applies some changes on top. The main thing was to
update path normalization to rely on Qt for mounts. This resolves a few
problems:

1. Target paths were not correctly processed on Windows: the previous
implementation used `std::filesystem` which would treat target paths as
Windows paths on that platform, whereas they are meant for Linux VMs. In
particular, that meant that a) paths were not correctly identified as
relative or absolute; b) they were not normalized correctly. Qt, on the
other hand, seems to understand Linux paths even on Windows. See
https://gist.github.com/ricab/312893d71ad09c6271107eb902f48fdd and
https://godbolt.org/z/j5f7657v4.

2. Target paths weren't fully normalized. In particular, trailing
slashes were not removed or added. The underlying call to
`std::filesystem::path::lexically_normal` kept `/foo/bar/` and
`/foo/bar` unchanged ¯\_(ツ)_/¯

It also streamlines the code around this a little and extracts a
constant for `/home/ubuntu`. We should actually have one for the
`ubuntu` user instead, but I didn't feel like replacing that all over in
here.

See commit messages for further details.

MULTI-1928