[tests] Fix sftp-related tests with premock
After the stfp_server_init change the libssh premock requires that a
new function be mocked in the tests and some adjustments are made. The
function to be mocked is sftp_reply_version, which requires minor
changes. The adjustment of sftp_get_client_message requires  more
adjustments due to how many times it is called in the codebase.
test_sftpserver.cpp already has a framework that simplifies the changes,
while test_sshfsmount.cpp requires the introduction of stateful lambdas
to make the tests possible.