[build] Fix Windows build with libssh internals
Unfortunately, it seems like MSVC requires deactivating the shared
libraries compilation in libssh to be able to access the functions
without the LIBSSH_API prefix, which in this case we are only interested
in sftp_reply_version, which used to happen internally inside
sftp_server_init. Seems like the current best solution, but still could
be suboptimal.