[build] Fix Windows build with libssh internals
Unfortunately, it seems like MSVC requires exposing all symbols to be
able to access the functions without the LIBSSH_API prefix, which in
this case we are only interested in sftp_reply_version, which used to
happen internally inside sftp_server_init. Seems like the current best
solution, but still could be suboptimal.

In this commit it is done by making the libssh library static, but later
a CMake option will be set to expose these symbols only on Windows.