Docs/fix windows build instructions (#4354)
Following from #4113, the build instructions for Windows were reviewed
and tested on Windows 11 Pro. The instructions were adapted to cover the
gap to the newer OS to allow building in its environment. The
installation of `git` and `qt6` were also unified to be done through
chocolatey and aqt respectively.

Additionally, general formatting and organization of the file was
changed minimally to improve visual clarity and to unify formatting in
the file.

The Windows 10 Pro specific instructions were left as-is when they would
conflict with the Windows 11 Pro instructions. In those cases, a
separate section for Windows 11 Pro was made.

Untested sections so far:
Permissions/privileges for multipassd in Windows 11
ninja package in Windows 11
cmder and powershell-only console setups in Windows 11
Installation while disabling the Anti-Virus in Windows 11