[cli-tests] Implement pre-post VM restart hook
Sometimes, the VM is not ready to accept connections after a "multipass
restart", causing subsequent checks to fail.

This patch introduces a pre and post hook to multipass() which is
triggered when the "restart" command is used. The pre() hook would grab
the boot-id's of the VMs that passed to the restart, and post() would
ensure that the multipass() won't return before the boot-id of each VM
has changed, ensuring both VM connectivity and the reboot.

Signed-off-by: Mustafa Kemal Gilor <mustafa.gilor@canonical.com>