[vcpkg-ports] Introduce gRPC superbuild wrapper
gRPC superbuild wrapper is a replacement for the previous patch file
we had for removing unneeded parts from gRPC build. The superbuild does
the same, but programmatically and in a less fragile manner.

This approach should make it way easier to upgrade gRPC since it's no
longer necessary to adapt a patch file.

Signed-off-by: Mustafa Kemal Gilor <mustafa.gilor@canonical.com>