[ip] Remove redundant "N/A" string in place of IPs
Stop adding an "N/A" string in place of a missing IP address for running
instances in list and info replies. I believe this was meant to transmit
that the IP was "Not Available" or "Not applicable" in the VirtualBox
case (in other backends, instances with no IP would have been in an
unknown state, not running). But VirtualBox is getting deprecated and
"N/A" isn't much clearer than "--" anyway. Furthermore, formatting
missing IPs is a presentation matter that should be left to clients.
Finally, the GUI will look better with a dash for both private and
public IPs (when they are both missing) than an N/A for private and a
dash for public.