[cmake] Build vcpkg release-only by default
The default `cmake configure` with CMAKE_BUILD_TYPE=DEBUG builds both
the debug and the release variants of the `vcpkg` deps due to how vcpkg
ecosystem is shaped (e.g. some deps need stuff that's only available in
the `release` configuration). This causes us to spend twice the time
waiting for the vcpkg to finish the building the deps, which is
especially troublesome in cold-builds (i.e. no cache).

This patch fixes that by unsetting the VCPKG_BUILD_DEFAULT parameter
in debug builds. VCPKG_BUILD_DEFAULT is now a option to the build.

To enable building debug variant of vcpkg deps, use the following:

> -DVCPKG_BUILD_DEFAULT:BOOL=ON

Signed-off-by: Mustafa Kemal Gilor <mustafa.gilor@canonical.com>