[cli-tests] Populate cfg as early as possible
Get rid of the store_config fixture in favor of pytest_configure.
request.addfinalizer with atexit.register to remove the request dep
from make_temporary_storage_dir, which in turn allows store_config
body to be moved to the pytest_configure.

The advantage of this approach is pytest_configure runs before the test
case discovery so cfg args can be used in the pytest decorators.

Signed-off-by: Mustafa Kemal Gilor <mustafa.gilor@canonical.com>