[macOS] Backwards compatible builds (#4496)
This PR adds the ability to build on the latest macOS runners while
still targeting an older version of macOS.

The `CMAKE_OSX_DEPLOYMENT_TARGET` is propagated to vcpkg so that all
dependencies are also built for the correct version of macOS.

Because OpenSSL was originally fetched via brew installation, its binary
would inherently target the version of macOS that it was being installed
on. To remedy this, OpenSSL is now fetched from vcpkg.

`windows-macos.yml` is subsequently pruned to no longer spawn unneeded
macOS runners.

The deployment target can be queried from a Multipass binary with the
command `otool -l multipassd | grep -A 3 LC_BUILD_VERSION`.

Updated licence information in the installer with the appropriate
licence as found in the [openssl
repository](https://github.com/openssl/openssl). Also, updated the
licence information for semver as its source recently changed in #4461.

Verified on a macOS 14 virtual machine. Nested virtualization is only
supported on M3 and newer, so you will need to remove the CPU checks in
`preinstall-multipassd.sh` and `CPack.distribution.dist.in`.

Closes https://github.com/canonical/multipass/pull/4023
Closes https://github.com/canonical/multipass-private/pull/739

---

MULTI-1920
MULTI-2272