Migrate QT to vcpkg (#4530)
Move QT to `vcpkg`. A couple of dependencies in Qt's dependency chain
(namely libsystemd, b2 and libxcrypt) needs a bunch of libraries to be
pre-installed so this PR also takes care of that, too.

Only the necessary parts of Qt are enabled and installed (core, network,
concurrent, openssl and dbus) to ensure minimum effect on the build
time, and to avoid introducing more Qt stuff to codebase accidentally.

Significant changes:
- We're now linking against Qt statically (default for all vcpkg deps).
- Had to drop the Qt version from 6.10 to 6.9 since it's the latest
available.
- Removed MacOS RPATH adjustment script for Qt
- Removed QT network information plugin code from CMake (the plugin
seems to be no longer in use)
- Removed instructions regarding manual Qt installation from the README
and build instructions